+++ statistics/table.go
@@ -29,9 +29,6 @@ import (
 	"github.com/pingcap/tidb/tablecodec"
 	"github.com/pingcap/tidb/types"
 	"github.com/pingcap/tidb/util/chunk"
-	"github.com/pingcap/tidb/util/codec"
-	"github.com/pingcap/tidb/util/collate"
-	"github.com/pingcap/tidb/util/mathutil"
 	"github.com/pingcap/tidb/util/ranger"
 	"go.uber.org/atomic"
 )
@@ -469,176 +466,6 @@ func (t *Table) IsOutdated() bool {
 		return true
 	}
 	return false
-}
-
-func (coll *HistColl) findAvailableStatsForCol(sctx sessionctx.Context, uniqueID int64) (isIndex bool, idx int64) {
-	// try to find available stats in column stats
-	if colStats, ok := coll.Columns[uniqueID]; ok && colStats != nil && !colStats.IsInvalid(sctx, coll.Pseudo) && colStats.IsFullLoad() {
-		return false, uniqueID
-	}
-	// try to find available stats in single column index stats (except for prefix index)
-	for idxStatsIdx, cols := range coll.Idx2ColumnIDs {
-		if len(cols) == 1 && cols[0] == uniqueID {
-			idxStats, ok := coll.Indices[idxStatsIdx]
-			if ok &&
-				idxStats.Info.Columns[0].Length == types.UnspecifiedLength &&
-				!idxStats.IsInvalid(sctx, coll.Pseudo) &&
-				idxStats.IsFullLoad() {
-				return true, idxStatsIdx
-			}
-		}
-	}
-	return false, -1
-}
-
-// GetSelectivityByFilter try to estimate selectivity of expressions by evaluate the expressions using TopN, Histogram buckets boundaries and NULL.
-// Currently, this method can only handle expressions involving a single column.
-func (coll *HistColl) GetSelectivityByFilter(sctx sessionctx.Context, filters []expression.Expression) (ok bool, selectivity float64, err error) {
-	// 1. Make sure the expressions
-	//   (1) are safe to be evaluated here,
-	//   (2) involve only one column,
-	//   (3) and this column is not a "new collation" string column so that we're able to restore values from the stats.
-	for _, filter := range filters {
-		if expression.IsMutableEffectsExpr(filter) {
-			return false, 0, nil
-		}
-	}
-	if expression.ContainCorrelatedColumn(filters) {
-		return false, 0, nil
-	}
-	cols := expression.ExtractColumnsFromExpressions(nil, filters, nil)
-	if len(cols) != 1 {
-		return false, 0, nil
-	}
-	col := cols[0]
-	tp := col.RetType
-	if types.IsString(tp.GetType()) && collate.NewCollationEnabled() && !collate.IsBinCollation(tp.GetCollate()) {
-		return false, 0, nil
-	}
-
-	// 2. Get the available stats, make sure it's a ver2 stats and get the needed data structure from it.
-	isIndex, i := coll.findAvailableStatsForCol(sctx, col.UniqueID)
-	if i < 0 {
-		return false, 0, nil
-	}
-	var statsVer, nullCnt int64
-	var histTotalCnt, totalCnt float64
-	var topnTotalCnt uint64
-	var hist *Histogram
-	var topn *TopN
-	if isIndex {
-		stats := coll.Indices[i]
-		statsVer = stats.StatsVer
-		hist = &stats.Histogram
-		nullCnt = hist.NullCount
-		topn = stats.TopN
-	} else {
-		stats := coll.Columns[i]
-		statsVer = stats.StatsVer
-		hist = &stats.Histogram
-		nullCnt = hist.NullCount
-		topn = stats.TopN
-	}
-	// Only in stats ver2, we can assume that: TopN + Histogram + NULL == All data
-	if statsVer != Version2 {
-		return false, 0, nil
-	}
-	topnTotalCnt = topn.TotalCount()
-	histTotalCnt = hist.NotNullCount()
-	totalCnt = float64(topnTotalCnt) + histTotalCnt + float64(nullCnt)
-
-	var topNSel, histSel, nullSel float64
-
-	// Prepare for evaluation.
-
-	// For execution, we use Column.Index instead of Column.UniqueID to locate a column.
-	// We have only one column here, so we set it to 0.
-	originalIndex := col.Index
-	col.Index = 0
-	defer func() {
-		// Restore the original Index to avoid unexpected situation.
-		col.Index = originalIndex
-	}()
-	topNLen := 0
-	histBucketsLen := hist.Len()
-	if topn != nil {
-		topNLen = len(topn.TopN)
-	}
-	c := chunk.NewChunkWithCapacity([]*types.FieldType{tp}, mathutil.Max(1, topNLen))
-	selected := make([]bool, 0, mathutil.Max(histBucketsLen, topNLen))
-
-	// 3. Calculate the TopN part selectivity.
-	// This stage is considered as the core functionality of this method, errors in this stage would make this entire method fail.
-	var topNSelectedCnt uint64
-	if topn != nil {
-		for _, item := range topn.TopN {
-			_, val, err := codec.DecodeOne(item.Encoded)
-			if err != nil {
-				return false, 0, err
-			}
-			c.AppendDatum(0, &val)
-		}
-		selected, err = expression.VectorizedFilter(sctx, filters, chunk.NewIterator4Chunk(c), selected)
-		if err != nil {
-			return false, 0, err
-		}
-		for i, isTrue := range selected {
-			if isTrue {
-				topNSelectedCnt += topn.TopN[i].Count
-			}
-		}
-	}
-	topNSel = float64(topNSelectedCnt) / totalCnt
-
-	// 4. Calculate the Histogram part selectivity.
-	// The buckets upper bounds and the Bucket.Repeat are used like the TopN above.
-	// The buckets lower bounds are used as random samples and are regarded equally.
-	if hist != nil && histTotalCnt > 0 {
-		selected = selected[:0]
-		selected, err = expression.VectorizedFilter(sctx, filters, chunk.NewIterator4Chunk(hist.Bounds), selected)
-		if err != nil {
-			return false, 0, err
-		}
-		var bucketRepeatTotalCnt, bucketRepeatSelectedCnt, lowerBoundMatchCnt int64
-		for i := range hist.Buckets {
-			bucketRepeatTotalCnt += hist.Buckets[i].Repeat
-			if len(selected) < 2*i {
-				// This should not happen, but we add this check for safety.
-				break
-			}
-			if selected[2*i] {
-				lowerBoundMatchCnt++
-			}
-			if selected[2*i+1] {
-				bucketRepeatSelectedCnt += hist.Buckets[i].Repeat
-			}
-		}
-		var lowerBoundsRatio, upperBoundsRatio, lowerBoundsSel, upperBoundsSel float64
-		upperBoundsRatio = mathutil.Min(float64(bucketRepeatTotalCnt)/histTotalCnt, 1)
-		lowerBoundsRatio = 1 - upperBoundsRatio
-		if bucketRepeatTotalCnt > 0 {
-			upperBoundsSel = float64(bucketRepeatSelectedCnt) / float64(bucketRepeatTotalCnt)
-		}
-		lowerBoundsSel = float64(lowerBoundMatchCnt) / float64(histBucketsLen)
-		histSel = lowerBoundsSel*lowerBoundsRatio + upperBoundsSel*upperBoundsRatio
-		histSel *= histTotalCnt / totalCnt
-	}
-
-	// 5. Calculate the NULL part selectivity.
-	// Errors in this staged would be returned, but would not make this entire method fail.
-	c.Reset()
-	c.AppendNull(0)
-	selected = selected[:0]
-	selected, err = expression.VectorizedFilter(sctx, filters, chunk.NewIterator4Chunk(c), selected)
-	if err != nil || len(selected) != 1 || !selected[0] {
-		nullSel = 0
-	} else {
-		nullSel = float64(nullCnt) / totalCnt
-	}
-
-	// 6. Get the final result.
-	res := topNSel + histSel + nullSel
-	return true, res, err
 }
 
 // PseudoAvgCountPerValue gets a pseudo average count if histogram not exists.
